Fixes #23 (#24)
* [Example] Update LifetimeTracker Pod version

* Add missing files to the library Xcode project and remove UIKit usage from non iOS targets

* Remove UIKit Dependency from localized Strings helper

* Remove UIKit dependency from Constants file

* [Example] Update LifetimeTracker Pod version

* [Example] Rename Example project to "Example" to avoid confusion to the actual LifetimeTracker project

* [Example] Add actual LifetimeTracker Xcode project to workspace to enable the editing of the library
