+++ Example/Example.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,10 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:LifetimeTracker.xcodeproj">
+      location = "group:Example.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:../LifetimeTracker.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
