+++ Sources/Constants.swift
@@ -3,6 +3,7 @@
 //  LifetimeTracker
 //
 //  Created by Hans Seiffert on 23.10.17.
+//  Copyright © 2017 LifetimeTracker. All rights reserved.
 //
 
 import Foundation
@@ -18,17 +19,6 @@ internal struct Constants {
 		struct Reuse {
 			static let dashboardCell = "lifetimeTrackerDashboardTableViewCell"
 			static let dashboardHeader = "lifetimeTrackerDashboardHeaderViewCell"
-		}
-	}
-
-	struct Layout {
-		static let animationDuration: TimeInterval = 0.3
-
-		struct Dashboard {
-			static let headerHeight: CGFloat = 44
-			static let cellHeight: CGFloat = 44
-			static let minTotalHeight: CGFloat = headerHeight + cellHeight
-			static let sectionHeaderHeight: CGFloat = 30
 		}
 	}
 }
